integrations/twilio/studio-flow-SMS/requirements.txt
gunicorn==19.9.0
Werkzeug==0.16.0
Flask==1.1.1
requests
Flask-Cors
google-cloud-dialogflow-cx==0.5.0